Fix compat issues
